moving CondFormats/BTagObject code into CondFormats/BTauObjects
