--- CondFormats/BTagObjects/BuildFile.xml
@@ -1,10 +1,0 @@
-<flags   GENREFLEX_ARGS="--"/>
-<use   name="CondFormats/Common"/>
-<use   name="FWCore/MessageLogger"/>
-<use   name="FWCore/Utilities"/>
-<use   name="boost"/>
-<use   name="rootrflx"/>
-<use   name="CondFormats/PhysicsToolsObjects"/>
-<export>
-  <lib   name="1"/>
-</export>
