--- CondFormats/BTagObjects/src/headers.h
@@ -1,3 +1,0 @@
-#include "CondFormats/BTagObjects/interface/BTagEntry.h"
-#include "CondFormats/BTagObjects/interface/BTagCalibration.h"
-#include "CondFormats/BTagObjects/interface/BTagCalibrationReader.h"
