--- CondFormats/BTagObjects/test/testSerializationBTagObjects.cpp
@@ -1,12 +1,0 @@
-#include "CondFormats/Serialization/interface/Test.h"
-
-#include "../src/headers.h"
-
-int main()
-{
-    testSerialization<BTagEntry>();
-    testSerialization<BTagEntry::Parameters>();
-    testSerialization<BTagCalibration>();
-
-    return 0;
-}
