+++ CondFormats/BTauObjects/src/T_EventSetup_BTagCalibration.cc
@@ -1,3 +1,3 @@
-#include "CondFormats/BTagObjects/interface/BTagCalibration.h"
+#include "CondFormats/BTauObjects/interface/BTagCalibration.h"
 #include "FWCore/Utilities/interface/typelookup.h"
 TYPELOOKUP_DATA_REG(BTagCalibration);
