+++ CondFormats/BTauObjects/src/T_EventSetup_BTagEntry.cc
@@ -1,3 +1,3 @@
-#include "CondFormats/BTagObjects/interface/BTagEntry.h"
+#include "CondFormats/BTauObjects/interface/BTagEntry.h"
 #include "FWCore/Utilities/interface/typelookup.h"
 TYPELOOKUP_DATA_REG(BTagEntry);
