+++ CondFormats/BTauObjects/src/headers.h
@@ -4,3 +4,7 @@
 #include "CondFormats/BTauObjects/interface/CombinedSVCategoryData.h"
 #include "CondFormats/BTauObjects/interface/CombinedSVCalibration.h"
 #include "CondFormats/BTauObjects/interface/CombinedTauTagCalibration.h"
+
+#include "CondFormats/BTauObjects/interface/BTagEntry.h"
+#include "CondFormats/BTauObjects/interface/BTagCalibration.h"
+#include "CondFormats/BTauObjects/interface/BTagCalibrationReader.h"
