+++ CondFormats/BTauObjects/test/BuildFile.xml
@@ -3,3 +3,15 @@
 <bin file="testSerializationBTauObjects.cpp">
     <use   name="CondFormats/BTauObjects"/>
 </bin>
+
+<bin file="testBTagEntry.cpp">
+    <use name="CondFormats/BTauObjects"/>
+</bin>
+
+<bin file="testBTagCalibration.cpp">
+    <use name="CondFormats/BTauObjects"/>
+</bin>
+
+<bin file="testBTagCalibrationReader.cpp">
+    <use name="CondFormats/BTauObjects"/>
+</bin>
